[repo view] Let "link with selection" be always enabled
It's actually a normal state toggling command for which a default
handler is fine. It didn't have a default handler because it was badly
implemented and assumed that the currently active part was the
repositories view.

Change the handling of all three toggle commands in the repositories
view. Let them only toggle their state, don't make them refresh the
view. Make the view listen to state changes and refresh if needed.
The content provider doesn't need to listen to changes; it has no way
to inform the viewer to refresh.

Also remove duplicating the state into local boolean variables; use
the state's current value instead.

Bug: 537958
Change-Id: I2027a64bd6f4e88439e7975b9d3c3eb3d381efef
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>