Close repositories when they are no longer referenced
EGit uses a WeakReference cache for Repository instances. Because
it has no real control over where a Repository instance is referenced
from and when exactly it would be safe to close a repository, it
just didn't do so. Once a repository really was no longer referenced,
it would be garbage collected and disappear from the central cache
in EGit. But even then the repository was not closed.

Fix this by making sure Repository instances are closed when garbage
collected. This required an overhaul of the RepositoryCache: with a
simple WeakReference<Repository> the repository is already garbage
collected before EGit gets notified about it. Instead introduce a
RepositoryHandle: this is a wrapper around a Repository and delegates
all calls to its underlying real Repository. The RepositoryCache only
hands out such RepositoryHandles. Use a ReferenceQueue and a daemon
Closer thread to manage WeakReferences.

When a handle is no longer referenced it will be garbage collected,
and its WeakReference will be put onto the ReferenceQueue. Keep a
reference to the real Repository in the WeakReference, though. That
way the real repository is still available and can be closed when
the Closer thread pops the WeakReference from the queue.

Because a RepositoryHandle behaves exactly like the real underlying
Repository, clients won't notice any difference. RepositoryEvents are
intercepted by the RepositoryHandle and by the RepositoryCache for the
global listener list, and are forwarded with the handle substituted
for the real Repository. Thus a client never has access to the
underlying Repository instance. The only caveat is that a client must
not register a listener on JGit's global listener list but instead use
the global listener list from EGit's RepositoryCache.

Bug: 548495
Change-Id: Idc41c4e71f800dbdbd21aa662a53a091f3692dcc
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>