[releng] egit.test.tmpdir not set in test runs in build
The egit.test.tmpdir variable is never set in maven builds. As a
result tests always use ~/egit.test.tmpdir as test directory. In
the CI build, that's outside the build workspace on a different
mount point.

Make sure that builds do use the directory specified in the CI job
script, which is a directory inside the build workspace. Define
a maven property and use it in ${coretest.vmparams} and in
${uitest.vmparams}.

Change-Id: I9f953167602cfa8382f2408b4944a3e509cf223f
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>