Fix handling of empty area with repo groups
The empty area should be shown only if indeed the CommonViewer is empty.
If there are repository groups but no repositories, the empty area
must _not_ be shown. On the other hand, if initially there is really
nothing and the empty area _is_ shown, the view must switch to the
viewer when a new repository group is created.

Change-Id: I0fc8fe2ea30faa326ce8d6e3b40b213700a7be0b
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>