[ref filter dialog] Prevent re-defining preconfigured filters
Otherwise these could suddenly be removed from the table. Add a
validation for this case.

Also ensure that an open cell editor is applied when any button
except "Cancel" is clicked in the dialog. Use a listener on the
ColumnViewerEditor to re-set the message and buttons when the cell
editor is deactivated; cancelEditor() or applyEditorValue() on the
listener on the cell editor itself are not always called.

Bug: 553449
Change-Id: I6df8cf16561911f1d1c8e6f8d7da4656ce461b0c
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>