CommitSelectionDialog: prevent InvalidObjectException and include HEAD
With general ref filters we may have refs pointing to non-commit
objects as start points of the walk, which would cause an exception.
Peel refs and generally be more careful with marking the start points
of the walk, like GitHistoryWalk. Also always include HEAD -- this
dialog is used for "Compare with Commit" and "Replace with Commit",
and it's highly confusing to get an empty history just because the
selected file isn't touched at all in the history as filtered by the
ref filters defined in the history page.

Bug: 553249
Change-Id: I3eac05b869e7c51dc6f33a7325f09b78f953fd1d
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>