Add Orbit version used in target platform to EGit repository
This ensures that EGit can be installed with only the EGit update
site being activated in Eclipse. If any dependencies are missing
locally in the EGit repository, p2 will try to fetch them from the
linked repositories, even if the user has unchecked "Check all update
sites" in the installation dialog.

Bug: 560438
Change-Id: I233d0422e2aa1da432c58b6b526a04f93728d049
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>