Staging View: Allow "Push HEAD" when there is nothing to commit
The "Commit and Push" button becomes "Push HEAD" in case there is
no change to be committed. This makes the most common push operation
directly accessible in the staging view and saves a few clicks in
some cases as the user doesn't have to switch to some other view.

If the repository has no HEAD (unborn branch) do not offer "Push
HEAD" but only "Commit and Push".

Other minor UI changes: don't require the buttons to be the same size
to avoid "jumping" buttons when the text on the "Commit and Push" or
"Push HEAD" button changes, and use the PushWizardDialog also for the
"Push To Gerrit..." wizard (so that the the "Finish" button is labeled
"Push"). Remove setHelpAvailable(true) calls on the dialog since there
is no help available. No point showing the help icon if it doesn't do
anything.

Bug: 512457
Change-Id: Ic67f17480458054bd8cdcf6b93061374da3b6c2f
Signed-off-by: Mickael Istria <mistria@redhat.com>
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>