Rename interactive rebase job
Interactive rebase uses job names corresponding to the jgit commands
implementing interactive rebase. The command processing all the steps
(and thereby doing the main part of the user visible operation) should
be named "Interactive rebase" instead to make it easy to recognize the
origin of that job in the progress view.

Bug: 562950
Change-Id: I881788a161766249082c0510d76dd30d80f3578a
Signed-off-by: Michael Keppler <Michael.Keppler@gmx.de>