Layout fix in RepositorySelectionPage
When configuring a push URI, the button to select a local bundle
file is not shown. However, the button to select a local directory
as upstream repository was set to span 3 columns in that case. That's
one too many, and broke the layout of the page completely in that
case.

Reduce the span to 2 columns: the one for the "Local Folder..."
button plus the one for the (not shown) "Local Bundle File..." button.

Change-Id: I2df6dc9bf466cd638b8aa0bdbcb9f3aff01c5c16
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>