Do not use the lightweight decorator in the staging view
This reverts commit a1a7ffd2c2467b4033b83307903a9bf54f560348.

Keep the simplifications in the StagingViewLabelProvider. Change the
ColorsAndFonts cache to track color and font changes itself (instead
of having the GitLightweightDecorator do so). Give the staging view's
label provider such a ChangeTrackingColorsAndFonts.

In the staging view, update buttons and messages only in response to
events from our own ProblemLabelDecorator; we're not interested in
other events there.

The staging view cannot use the GitLightweightDecorator because it
doesn't decorate entries for non-workspace files, and making it do so
leads to double decorations on entries that _do_ have resources. Also,
going via IResource when a StagingEntry has all the needed information
ready is less efficient. For staging entries, decorating on the UI
thread is actually feasible.

Bug: 565990
Change-Id: Ic32682334884a75a6e52a33039dc72cac464e0fd
Signed-off-by: Thomas Wolf <thomas.wolf@paranor.ch>